pyproject.toml
[project]
name = "healthcare-graph-agent"
version = "2.0.0"
description = "Modular healthcare analytics platform with AI agent and graph database integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Justin Martin"}
]
keywords = ["healthcare", "referral-network", "graph-database", "cosmos-db", "ai-agent"]

dependencies = [
    "gremlinpython>=3.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
azure = [
    "azure-functions",
    "openai>=1.0.0",
]
gradient = [
    "gradient-adk>=0.1.9",
    "gradient>=3.10.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.mypy_cache
    | powerbi_export
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "__pycache__", "powerbi_export"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["powerbi_export", ".venv"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*", "scripts*", "config*"]
